+++ packages/renterd/lib/src/abstract/host_absts/get_block_list_abst.dart
@@ -1,0 +1,5 @@
+import 'package:http/http.dart' as http;
+
+abstract class GetBlockListAbst {
+  Future<http.Response> getBlockList();
+}