+++ packages/renterd/lib/src/abstract/host_absts/get_host_info_by_public_key_abst.dart
@@ -1,0 +1,5 @@
+import 'package:http/http.dart' as http;
+
+abstract class GetHostInfoByPublicKeyAbst {
+  Future<http.Response> getHostInfoByPublicKey({required String publicKey});
+}