+++ packages/renterd/lib/src/abstract/host_absts/get_host_scanning_abst.dart
@@ -1,0 +1,9 @@
+import 'package:http/http.dart' as http;
+
+abstract class GetHostScannigAbst {
+  Future<http.Response> getHostScannig({
+    required int offset,
+    required int limit,
+    required String lastScan,
+  });
+}