+++ packages/renterd/lib/src/abstract/host_absts/remove_hosts_abst.dart
@@ -1,0 +1,8 @@
+import 'package:http/http.dart' as http;
+
+abstract class RemoveHostsAbst {
+  Future<http.Response> removeHosts({
+    required int minRecentScanFailures,
+    required String maxDowntimeHours,
+  });
+}