+++ packages/renterd/lib/src/abstract/host_absts/update_allow_list_abst.dart
@@ -1,0 +1,8 @@
+import 'package:http/http.dart' as http;
+
+abstract class UpdateAllowListAbst {
+  Future<http.Response> updateAllowList({
+    List<String>? addHostList,
+    List<String>? removeHostList,
+  });
+}