+++ packages/renterd/lib/src/abstract/host_absts/update_block_list_abst.dart
@@ -1,0 +1,8 @@
+import 'package:http/http.dart' as http;
+
+abstract class UpdateBlockListAbst {
+  Future<http.Response> updateBlockList({
+    List<String>? addHostList,
+    List<String>? removeHostList,
+  });
+}