+++ packages/renterd/lib/src/injection.dart
@@ -12,6 +12,15 @@ import 'abstract/accounts_absts/update_balance_abst.dart';
 import 'abstract/consensus_absts/accept_block_abst.dart';
 import 'abstract/consensus_absts/get_sia_fund_fee_abst.dart';
 import 'abstract/consensus_absts/get_state_abst.dart';
+import 'abstract/host_absts/get_allow_list_abst.dart';
+import 'abstract/host_absts/get_block_list_abst.dart';
+import 'abstract/host_absts/get_host_info_by_public_key_abst.dart';
+import 'abstract/host_absts/get_host_scanning_abst.dart';
+import 'abstract/host_absts/get_hosts_abst.dart';
+import 'abstract/host_absts/post_interaction_abst.dart';
+import 'abstract/host_absts/remove_hosts_abst.dart';
+import 'abstract/host_absts/update_allow_list_abst.dart';
+import 'abstract/host_absts/update_block_list_abst.dart';
 import 'data/services/accounts_services/add_deposit_impl.dart';
 import 'data/services/accounts_services/get_all_accounts_impl.dart';
 import 'data/services/accounts_services/get_an_account_by_id_impl.dart';
@@ -23,6 +32,15 @@ import 'data/services/accounts_services/update_balance_impl.dart';
 import 'data/services/consensus_services/accept_block_impl.dart';
 import 'data/services/consensus_services/get_sia_fund_fee_impl.dart';
 import 'data/services/consensus_services/get_state_impl.dart';
+import 'data/services/host_services/get_allow_list_impl.dart';
+import 'data/services/host_services/get_block_list_impl.dart';
+import 'data/services/host_services/get_host_info_by_public_key_impl.dart';
+import 'data/services/host_services/get_host_scanning_impl.dart';
+import 'data/services/host_services/get_hosts_impl.dart';
+import 'data/services/host_services/post_interaction_impl.dart';
+import 'data/services/host_services/remove_hosts_impl.dart';
+import 'data/services/host_services/update_allow_list_impl.dart';
+import 'data/services/host_services/update_block_list_impl.dart';
 import 'logic/controllers/accounts_controllers/add_deposit_controller.dart';
 import 'logic/controllers/accounts_controllers/get_all_accounts_controller.dart';
 import 'logic/controllers/accounts_controllers/get_an_account_by_id_controller.dart';
@@ -34,6 +52,15 @@ import 'logic/controllers/accounts_controllers/update_balance_controller.dart';
 import 'logic/controllers/consensus_controllers/accept_block_controller.dart';
 import 'logic/controllers/consensus_controllers/get_sia_fund_fee_controller.dart';
 import 'logic/controllers/consensus_controllers/get_state_controller.dart';
+import 'logic/controllers/host_controllers/get_allow_list_controller.dart';
+import 'logic/controllers/host_controllers/get_block_list_controller.dart';
+import 'logic/controllers/host_controllers/get_host_info_by_public_key_controller.dart';
+import 'logic/controllers/host_controllers/get_host_scanning_controller.dart';
+import 'logic/controllers/host_controllers/get_hosts_controller.dart';
+import 'logic/controllers/host_controllers/post_interaction_controller.dart';
+import 'logic/controllers/host_controllers/remove_hosts_controller.dart';
+import 'logic/controllers/host_controllers/update_allow_list_controller.dart';
+import 'logic/controllers/host_controllers/update_block_list_controller.dart';
 
 final sl = GetIt.instance;
 
@@ -51,6 +78,16 @@ Future<void> initialization() async {
   sl.registerLazySingleton<AcceptBlockAbst>(() => AcceptBlockImpl());
   sl.registerLazySingleton<GetStateAbst>(() => GetStateImpl());
   sl.registerLazySingleton<GetSiaFundFeeAbst>(() => GetSiaFundFeeImpl());
+  sl.registerLazySingleton<GetHostInfoByPublicKeyAbst>(
+      () => GetHostInfoByPublicKeyImpl());
+  sl.registerLazySingleton<GetHostsAbst>(() => GetHostsImpl());
+  sl.registerLazySingleton<UpdateAllowListAbst>(() => UpdateAllowListImpl());
+  sl.registerLazySingleton<UpdateBlockListAbst>(() => UpdateBlockListImpl());
+  sl.registerLazySingleton<GetAllowListAbst>(() => GetAllowListImpl());
+  sl.registerLazySingleton<GetBlockListAbst>(() => GetBlockListImpl());
+  sl.registerLazySingleton<PostInteractionAbst>(() => PostInteractionImpl());
+  sl.registerLazySingleton<RemoveHostsAbst>(() => RemoveHostsImpl());
+  sl.registerLazySingleton<GetHostScannigAbst>(() => GetHostScannigImpl());
 
 //! Controller
 
@@ -76,4 +113,32 @@ Future<void> initialization() async {
       () => GetStateController(getStateAbst: sl.call()));
   sl.registerFactory<GetSiaFundFeeController>(
       () => GetSiaFundFeeController(getSiaFundFeeAbst: sl.call()));
+  sl.registerFactory<GetHostInfoByPublicKeyController>(
+      () => GetHostInfoByPublicKeyController(
+            getHostInfoByPublicKeyAbst: sl.call(),
+          ));
+  sl.registerFactory<GetHostsController>(() => GetHostsController(
+        getHostsAbst: sl.call(),
+      ));
+  sl.registerFactory<UpdateAllowListController>(() => UpdateAllowListController(
+        updateAllowListAbst: sl.call(),
+      ));
+  sl.registerFactory<UpdateBlockListController>(() => UpdateBlockListController(
+        updateBlockListAbst: sl.call(),
+      ));
+  sl.registerFactory<GetAllowListController>(() => GetAllowListController(
+        getAllowListAbst: sl.call(),
+      ));
+  sl.registerFactory<GetBlockListController>(() => GetBlockListController(
+        getBlockListAbst: sl.call(),
+      ));
+  sl.registerFactory<PostInteractionController>(() => PostInteractionController(
+        postInteractionAbst: sl.call(),
+      ));
+  sl.registerFactory<RemoveHostsController>(() => RemoveHostsController(
+        removeHostsAbst: sl.call(),
+      ));
+  sl.registerFactory<GetHostScannigController>(() => GetHostScannigController(
+        getHostScannigAbst: sl.call(),
+      ));
 }