+++ packages/renterd/lib/src/logic/controllers/host_controllers/get_hosts_controller.dart
@@ -1,0 +1,15 @@
+import 'package:http/http.dart' as http;
+
+import '../../../abstract/host_absts/get_hosts_abst.dart';
+
+class GetHostsController {
+  final GetHostsAbst getHostsAbst;
+
+  GetHostsController({
+    required this.getHostsAbst,
+  });
+
+  Future<http.Response> call() async {
+    return await getHostsAbst.getHosts();
+  }
+}