+++ packages/renterd/CHANGELOG.md
@@ -3,6 +3,16 @@
   - [x] Accept Block
   - [x] Get Sia Fund Fee
   - [x] Get State
+### Hosts Services Created
+  - [x] Get Allow list
+  - [x] Get Block list
+  - [x] Get Host Info By Public Key
+  - [x] Get Hose Scanning
+  - [x] Get Hosts
+  - [x] Post Interaction
+  - [x] Remove Hosts
+  - [x] Update Allow List
+  - [x] Update Block List
 ---------------------------------
 
 ## 0.0.1